GUI/MemWatch: fix a buffer overrun as well as initialize the DlgAddWatch correctly
This bug took valgrind to find and as of this commit, valgrind is better than bread and butter :)

On a serious note, this fixes a set of random crashes that could occur when one entry has a variable length type.  It also allowed to discover some weird initialization bugs in the add watch dialogs which are now fixed in this commit.