GUI/MemWatch: Signal an editing is in progress for ALL edits
This fixes a bug where the labels are constantly refreshed if there's at least one group contains a children.  This was caused by 2 problems: the update recursive function was emitting data changed on groups which doesn't make sense and also because Qt wants to updates all the data of the row if you emit a data changed which means it is not only going to refresh the value, but also the label.

This is fixed by only emitting when the node is not a group and always put the global editing flag to true/false for ALL edits including labels.