Update .gitignore with more files to ignore
New comment (after rebasing):
This keeps the minimal changes, only adding a rule for Sources/bin and
Sources/GeneratedFiles.

Old comment:
Mainly for the /bin and /obj directories that the VS project writes to.
The reason for not having those in the Source directory is so that the
solution explorer stays clean.
The new rules assume that no /Build, /Binary, /obj and /bin folders will
be created.

squash