Add MsBuild target to copy required files
This create a new target that gets executed after the build has finished. The
target uses a list of files (with optional sub directory) to read from and copy
to. The debug and release files are packed together, since the only difference
is a `d` at the end of the filenames.
This also supports DLLs for both configurations, by omitting the suffix.

The target is present in the project files - I don't think a small target like
this requires a new file. When we add more, it might be worth creating.