Remove MFP casts from connect calls
The static_casts are bloating the code, making it unnecessarily hard to read. The templated connect functions from Qt are smart enough to find the correct MFP.
Only when a signal has multiple overloaded possiblilities, the compiler will have cannot select one, but for this Qt has a template QOverload<>, in which you only have to provide the argument types to let it pick the correct one. Using this gets rid of the return type and duplicated class name in those calls.
C++14 has an even more elegant solution, but that failed to compile on my end.