DolphinProcess: Add support for the experimental Qt front-end
As this will eventually be the default in Dolphin, when that happens, a commit that adjust it will be done.  When 6.0 will come out with no Wx front-end, a commit will be done to only use the name of the process as on all version, it basically uses the same name for the default front-end.